Feature : screen refresh

+++ include/termsize.h
@@ -1,10 +1,12 @@
 #pragma once
 
+#include <stdint.h>
+
 struct termsize {
-	unsigned int cols;
-	unsigned int rows;
-	unsigned int xpix;
-	unsigned int ypix;
+	uint16_t cols;
+	uint16_t rows;
+	uint16_t xpix;
+	uint16_t ypix;
 };
 
 void ref_termsize(struct termsize* ref);
